projectile variable assigned

+++ UFO Game/Assets/_Completed-Assets/Scripts/ProjectileController.cs
@@ -7,7 +7,7 @@ public class ProjectileController : MonoBehaviour
     public float velocity = 3f;
     private float lifespan = 5f;
     private float startTime = 0f;
-    public float playerdamage;
+    public int playerdamage;
     Rigidbody2D rb2d;
     // Start is called before the first frame update
     void Start()
